% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rrm_datasets.R
\name{clean_rrm_datasets}
\alias{clean_rrm_datasets}
\title{Clean RRM Datasets}
\usage{
clean_rrm_datasets(
  database,
  hh_number,
  ind_number,
  hhsize,
  start_date,
  end_date
)
}
\arguments{
\item{database}{A dataframe containing the dataset to be cleaned.}

\item{hh_number}{A numeric column representing the household count.}

\item{ind_number}{A numeric column representing the number of individuals.}

\item{hhsize}{An integer value representing the estimated average household size from MSNA.}

\item{start_date}{A column representing the start date (eg:incident date; response start date).}

\item{end_date}{A column representing the end date (eg: validation date; response end date).}
}
\value{
A cleaned dataframe with corrected date formats, household and individual counts, and imputed dates.
}
\description{
This function applies multiple cleaning steps to an RRM dataset:
\itemize{
\item Converts \code{start_date} and \code{end_date} to Date format.
\item Ensures household (\code{hh_number}) and individual (\code{ind_number}) counts are logically consistent.
\item Handles missing dates and removes outliers based on the IQR method.
}
}
\examples{
library(dplyr)

data <- data.frame(
  hh_number = c(10, NA, 15, 50, NA, 30),
  ind_number = c(50, 30, NA, 100, 60, 20),
  start_date = c("2024-01-01", NA, "2024-03-01", "2024-04-01", NA,NA),
  end_date = c("2024-01-10", "2024-02-15", NA, "2024-05-01", "2024-06-15",NA)
)

cleaned_data <- clean_rrm_datasets(data,
                                   hh_number = "hh_number",
                                   ind_number = "ind_number",
                                   hhsize = 5,
                                   start_date = "start_date",
                                   end_date = "end_date")

print(cleaned_data)

}
