% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_monthly_summary.R
\name{compute_monthly_summary}
\alias{compute_monthly_summary}
\title{Compute Monthly Summary of Responses}
\usage{
compute_monthly_summary(
  response_data,
  processed_response_data,
  uuid,
  prev_period_date,
  current_period_date,
  start_response_date,
  ind_number,
  response_number,
  desag_by = NA
)
}
\arguments{
\item{response_data}{A dataframe containing response data.}

\item{processed_response_data}{A dataframe containing processed response data with
deduplicated alerts.}

\item{uuid}{The column name (as a string) representing unique identifiers for each response.}

\item{prev_period_date}{The start date of the previous period.}

\item{current_period_date}{The start date of the current period.}

\item{start_response_date}{The column name (as a string) representing the start date of a response.}

\item{ind_number}{The column name (as a string) representing the number of individuals
reached by a response.}

\item{response_number}{The column name (as a string) representing the number of responses.}

\item{desag_by}{(Optional) The column name (as a string) by which to disaggregate the
summary (e.g., region, sector). Default is \code{NA}.}
}
\value{
A dataframe summarizing the number of responses initiated (\code{response_started}),
the number of alerts that received at least one response (\code{alerts_with_at_least_oneresponse}),
and the total number of individuals reached (\code{people_reached}). If \code{desag_by} is
provided, results are grouped accordingly.
}
\description{
This function calculates a monthly summary of responses based on given response
data and processed response data. It computes the total number of responses
initiated, unique alerts with at least one response, and the number of individuals reached.
}
\examples{
library(dplyr)

# Sample response data
response_data <- data.frame(
  uuid = c("A1", "A1", "B2", "B2", "C1"),
  start_response_date = as.Date(c("2024-01-05", "2024-01-10", "2024-02-15",
  "2024-01-20", "2024-03-01")),
  ind_number = c(10, 20, 15, 30, 25),
  region = c("North", "North", "South", "South", "East"),
  donor=c("a1","a2","a1","a1","a3")
  )

# Sample processed response data (deduplicated alerts)
processed_response_data <- data.frame(
uuid = c("A1", "B2", "C1"),
ind_number = c(100, 50, 75),
response_number = c(2, 2, 1),
region = c("North", "South", "East")
)

# Compute summary with disaggregation by region
compute_monthly_summary(
  response_data = response_data,
  processed_response_data = processed_response_data,
  uuid = "uuid",
  prev_period_date = as.Date("2024-01-01"),
  current_period_date = as.Date("2024-02-01"),
  start_response_date = "start_response_date",
  ind_number = "ind_number",
  response_number = "response_number",
  desag_by = "region"
)
}
