% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_multiple_choice_sector_variables.R
\name{discretize_multiple_choice_variables}
\alias{discretize_multiple_choice_variables}
\title{Discretize Multiple Choice Variables}
\usage{
discretize_multiple_choice_variables(df, multi_choice_column, category_labels)
}
\arguments{
\item{df}{A dataframe containing multiple-choice responses.}

\item{multi_choice_column}{The name of the column containing multiple-choice responses (comma-separated string).}

\item{category_labels}{A named character vector where names represent the new column names
and values represent the corresponding label to search for in the \code{multi_choice_column}.}
}
\value{
A dataframe with additional binary indicator columns for each specified category.
}
\description{
This function scans a specified column in a dataset to generate binary indicators
for multiple-choice responses. It checks for the presence of predefined categories
and creates new columns representing whether each category is present (\code{1}) or absent (\code{0}).
}
\examples{
library(dplyr)
library(stringr)

df <- data.frame(
  uuid = c("A1", "B2", "C3"),
  selected_options = c("Food, WASH, NFI", "Health, Education, Protection", "Shelter, Livelihoods")
)

category_labels <- c(
  food = "Food",
  wash = "WASH",
  nfi = "NFI",
  shelter = "Shelter",
  health = "Health",
  education = "Education",
  protection = "Protection",
  livelihoods = "Livelihoods",
  fortified_flour = "Fortified Flour",
  mhm = "MHM"
)

df <- discretize_multiple_choice_variables(df, "selected_options", category_labels)
print(df)
}
