% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_date.R
\name{format_date}
\alias{format_date}
\title{Format Date Columns}
\usage{
format_date(database, dates_vector)
}
\arguments{
\item{database}{A dataframe containing date columns.}

\item{dates_vector}{A character vector of column names to convert to Date format.}
}
\value{
A dataframe with formatted date columns.
}
\description{
This function converts specified columns in a dataframe to Date format.
}
\examples{
df<-data.frame(
date1=c("2025-01-28", "2025-01-27", "2025-01-26"),
organisation_name = c("org_1", "org_2", "org_3"),
date2=c("2025-02-05", "2025-01-30", "2025-02-05")
)

df<-format_date(database=df,dates_vector=c("date1","date2"))

class(df$date1)
class(df$date2)
}
