% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotwider_freq_table.R
\name{pivotwider_freq_table}
\alias{pivotwider_freq_table}
\title{Reformater un tableau de frequences verticale en horizontale}
\usage{
pivotwider_freq_table(
  df,
  admin,
  pop_group,
  question,
  reponse,
  value,
  moe_var,
  n_var
)
}
\arguments{
\item{df}{Dataframe d'entrée}

\item{admin}{Variable d'administration (symbole non quoté)}

\item{pop_group}{Variable du groupe de population (symbole non quoté)}

\item{question}{Variable contenant la question (symbole non quoté)}

\item{reponse}{Variable contenant la réponse (symbole non quoté)}

\item{value}{Variable contenant la valeur principale (symbole non quoté)}

\item{moe_var}{Variable contenant la marge d'erreur (symbole non quoté)}

\item{n_var}{Variable contenant le nombre d'observations (symbole non quoté)}
}
\value{
Un dataframe résumé avec les valeurs pivotées et les statistiques moe/n
}
\description{
Cette fonction pivote les données, calcule les statistiques (moyennes, médianes,
et effectifs) pour les variables de type marge d'erreur et n, puis retourne un
tableau résumé par groupes. Elle se base sur les sorties d'analyse de IMPACT Initiatives
}
\examples{
# pivotwider_freq_table(df, admin = region, pop_group = group,
#                  question = q, reponse = r, value = val,
#                  moe_var = moe, n_var = n)
}
